requirements.txt
demoji==1.1.0
gdown==4.6.4
matplotlib==3.6.3
nerd==1.0.0
entity-fishing-client
nltk==3.7
numpy
pandas==1.5.3
python_Levenshtein==0.20.9
regex==2022.10.31
requests==2.25.1
scikit_learn==1.2.2
scipy==1.10.1
sentence_transformers==2.2.2
tensorflow_hub==0.12.0
torch==1.11.0
trectools==0.0.49
tensorflow==2.11.0
spacy
